backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
eventlet==0.33.3
SQLAlchemy==2.0.25
Werkzeug==2.3.8